requirements.txt
Django
django-filter==24.3